fix incorrect too-tight assertion, update tests and debuglns
It isn't correct to assert that the nextlink is never arbitrary data -- only a valid link. It is correct only to assert that *if* you successfully pop that nextlink into place as the new head, then it contained a valid link. This corrects those assertions.

Also add debugln's (commented-out) and more tests.