smalloc version 4!
Notable differences from version 3:

1. There are now 26 different size classes in the "large" sizeclass category so
that you can request very large allocations from smalloc. At the same time,
smalloc no longer falls back to the system allocator when it is exhausted --
instead returns a null pointer. Reason: I had a lot of trouble with the system
allocator on Macos, problems which I was never able to debug, and anyway with
the new larger sizeclasses, exhaustion will happen only in very special cases.
Finally, doing away with the fall-back-to-sys-allocator nicely eliminates some
code.

2. In Smalloc v3 there were two variables associated with each slab -- a
free-list-head ("flh") which pointed to the first-on-the-free-list free slot,
or had a 0 in it if the free list was empty, and an ever-allocated-counter,
which contained the number of slots that had ever been allocated. Therefore
each slot had three possible states: never-yet-allocated (it started in this
state), allocated, and freed. Each slot, when it was in the freed state,
contained the slot number of the next slot in the free-list.

This design was, I guess, a kind of "zero is initialization".

https://blog.xoria.org/zii/

But smalloc v4 has a really cool improvement! Slots which are in the freed
state contain, not the slot number of the next-in-free-list slot, but instead
an *encoding* of that slot number such that when it is zero then it decodes to
the next higher slot number. Therefore when a slot contains zeros, it points to
the numerically-next-slot-number as the next-in-the-free-list slot. This means
that when the entire slab and flh are all initialized to 0, they contain a
complete linked list of all the slots.

This means we do not need an ever-allocated-counter anymore. Now slots have
only two states: freed (they start in this state), and allocated.

This design is even *more* a kind of "zero is initialization", I guess. :-)

But the part I like best about it is that it eliminates one of the two
variables and eliminates some code. A downside is that it does require just a
*few* more instructions inside the innermost code to push or pop from the free
list. The extra instructions are for encoding and respectively decoding the
"next-free-list-entry-link".

3. The locations of all things (the smalloc base pointer, and all slots, slabs,
and flhs) are now lined up with nice power-of-two boundaries so that we can use
bit-twiddling instead of integer arithmetic to map from pointers to slabs,
slots, and flhs.

Also the sizeclasses themselves are lined up in such a way that given a
pointer, you can compute most of the information about which sizeclass the
pointer falls into by just a count-leading-zeros instruction. 😃

Reason: It was fun!

Also, I hope the bit-twiddling will result in shorter asm code and faster
runtimes, but I haven't actually tested that yet. On the downside, this makes
the code *much* much harder to read and understand. This effect is so
pronounced that I'm considering redefining "smalloc" to mean "sparse malloc"
instead of "simple malloc". 😭

STATUS OF VERSION 4:

It passes all 80 unit tests.

There is quite a bit of mess to be cleaned up:

1. The benchmarks are broken because I gave up on Criterion and haven't yet
tried a replacement. I intend to try Tango soon.

2. I think maybe some other unit tests are commented-out? Imma go check for
that.

3. The README.md, which I lovingly maintained from pre-smalloc-v0 up until
smalloc-v3 is currently broken because I haven't updated it for smalloc-v4.

4. There are quite a few parts of the code that I've been intending to return
to and clean up. There are a whole lot of commented-out debug prints, and
comments saying things like "//xxx come back and see about changing this in a
certain way".