smalloc v5.0.0!
Move the flhs from being stored in the last slot (or the last two slots in the
case of sizeclass 0) to being in another section. This allows us to remove the
code for special-casing sizeclass 0, and improves locality of accesses to the
flhs.

Also, cut in half all of the slabs (1 fewer bits' worth of slots) to avoid a
problem where when Tango tries to load it, the mmap fails. I don't understand
why that happened, but this change does indeed make it stop happening, and I
figure there may be other situations where using up less of the virtual memory
space could be good.

Create some very basic benchmarking -- see src/bin/simplebench.rs (standalone
executable) and benches/smallocb.rs (for use in the Tango benchmarking tool).

Make more of the constant values be symbolicly computed at compile-time instead
of hard-coded.