smalloc v6!
It is much simpler in terms of lines of code, even compared to smalloc v5!

It passes all tests.

Benchmarks indicate that it is faster than builtin global alloc even in the
heavy-multithreading case, which confuses me because previous benchmarks (way
back at like version 4 or so) seemed to indicate the opposite, and when Marek
and I tried smalloc v5 in Zebra in San Diego that seemed to slow Zebra down to
1/2 speed, which was consistent with those previous benchmarks. So this needs
further investigation.

Oh, but actually this version (v6.0.0) has Andrew Reece's suggestion of
per-thread-num slabs for all size classes, so that would explain why this is
faster than v4 and v5! Yay!

It doesn't have overflow handling! Need to put that back. Also it doesn't have
fallback-to-mmap anymore, and I need to put that back, too. And the README.md
is totally obsolete -- hasn't been updated since v4 or so.

But it is soooo few lines of code! 380 lines of code! Whoo-hoo!