three changes:
1. allocate space for sizeclass 0 and 1 but never use that space.

This allows us to eliminate "sc + 2" to convert to slotsizebits, and in general
eliminate notions of slotsizebits as a separate notion from sc. It also allows
us to put the flhs in the unused sizeclass 0 space, eliminating the notion of
the flhsbp.

2. move the bulk of `alloc` (excluding the translation from layout to sc) into
`inner_alloc`, and call `inner_alloc` from `alloc`, from the overflow-handling
in `inner_alloc`, and from `realloc`. This eliminates unnecessary translation
between Layout and sc and vice versa.

3. a new implementation of req_to_sc (which replaces req_to_slotsizebits) which
is delightfully short and sweet in both Rust and asm