add benchmark of the "hotspot" behavior were there are a lot of threads hammering on one particular flh and no threads doing anything else
simplify the benchmark macros (thanks, Claude Opus 4.5!)

note that the "63 cool threads" constant is designed to maximize the contention
on a single flh for smalloc. so this benchmark is in that sense "unfair" to
smalloc when using this benchmark to compare against other allocators.
Presumably there is a different constant number of cool threads that would
maximize contention in mimalloc, and a different one that would maximize
contention in jemalloc, etc. Or maybe not because maybe they dynamically adapt
right away. In any case, smalloc is currently very slow at this particular
benchmark, as expected (see below), and I'm intending to make it dynamically
adapt the thread-to-slab mapping next.