put back in the Growers policy
I'm still not sure if this is a good idea. The simplebench microbenchmarks
don't show any appreciable benefit (or cost) of having it in. The simd_json
benchmarks likewise.

And it might be unnecessary to do within smalloc because other layers both
above and below try to do the same thing. For example, the Rust vector class
will double the capacity if you extend a vector beyond its current capacity.

But the main things that keep me coming back to it are:

1. The memory-profiling I did on Zcash zebra showed that it was extending a
vector over and over up until 2 MiB (and in a few cases 3 MiB and 4 MiB), and
the default behavior of the Rust vector and the underlying allocator was
copying that data over and over -- trillions of bytes worth of memcpy in the
experiments I ran.

2. And this is unnecessary, if you think about the behavior of the virtual
memory subsystem. If you extend an allocation from 16 KiB to 17 KiB, and the
underlying memory allocator can't extend the underlying allocation, so you are
going to have to memcpy the 16 KiB of data to a new location, then you might as
well allocate a 32 GiB slot for the new location. This will spare you
subsequent memcpy's if the vector grows again (such as happens in Zcash zebra,
and I would guess also a lot of Rust codebases), and it doesn't cost anything!
The virtual memory subsystem is going to manage the memory in 4 KiB pages (or
16 KiB on Apple OSes) anyway, so allocating a 32 GiB slot doesn't cost any more
than allocating a 32 KiB slot.

So the Growers rule is back in, for now, but this is provisional depending on
some solid benchmarks (ideally in vivo -- a real-world usage in action) or
future reconsideration.

P.S. Well, actually, in current smalloc, it *does* cost something, which is
it increases the possibility of exhaustion due to all sufficiently large slots
being allocated. The current version of the Growers policy tries to mitigate
that issue by upgrading a grower to first a 256 KiB slot, and from there to a 2
MiB slot, because in current smalloc there are only around 1,000,000 total
slots of size >= 2 MiB. Another benefit of stopping over at 256 KiB before
jumping to 2 MiB in this version of the Growers policy is that if the system is
using large/huge/super-pages, we can pack eight 256 KiB allocations into a single
page, thus reducing demand for paging.

This is one of the things that could be improved if smalloc could rely on
5-level-page-tables everywhere, then we could have enough large slots that the
Growers rule could promote growers directly, once they are too big to fit
multiple into a memory page, to some huge size allocation. But unfortunately
Apple OSes do not yet support 5-level-page-tables.