core: fixed the performance regression! I had swapped the position of slabnum bits and sizeclass bits in the flhptr, which caused destructive interference ("false sharing") of cache lines. Claude helped me figure it out.
Funny: this is the *third* time I've encountered this. I did it right
originally, then when Andrew Reece suggested that I might have done it the
other way and gotten destructive interference I forgot I had done it right the
first time, then when doing this port I actually *did* do it the wrong way and
Claude helped me figure it out this time.

Also: simplify the failover logic drastically, to just "+= 7%32". To be
benchmarked.