core: remove growers strategy
It doesn't help with the simd-json benchmarks on either Apple M4 Max or Intel
Xeon Gold 3152. On the other hand, it doesn't hurt, either. (I suppose
simd-json just doesn't extend allocations a lot.)

It might help with Zebra. If I get around to benchmarking with Zebra again I'll
put it back in.